requirements.txt
pillow
boto3
pandas